+++ pos-system/src/main/java/com/lifepill/possystem/dto/requestDTO/RequestOrderSaveDTO.java
@@ -14,7 +14,7 @@ import java.util.Set;
 @Data
 public class RequestOrderSaveDTO {
 
-    private int cashiers;
+    private long cashiers;
     private Date orderDate;
     private Double total;
     private List<RequestOrderDetailsSaveDTO> orderDetails;
